backend/requirements.txt
fastapi
uvicorn[standard]
transformers>=4.0.0
torch
pyyaml
pytest
pytest-asyncio
python-dotenv
